+++ imports/api/branch/branchCollection.js
@@ -1,0 +1,3 @@
+import { Mongo } from 'meteor/mongo';
+
+export const Branches = new Mongo.Collection('Branches');