+++ imports/api/customers/customerMethods.tests.js
@@ -19,19 +19,12 @@ if (Meteor.isServer) {
     afterEach(() => {
       StubCollections.restore();
     });
-    
+
     describe('customer.new', () => {
       it('insert a new customer into the collection', () => {
         Customers.find({}).count().should.equal(0);
         Meteor.call('customer.new', {});
         Customers.find({}).count().should.equal(1);
-      });
-
-      it('only accepts an object as a parameter', () => {
-        (() => Meteor.call('customer.new', 1)).should.throw();
-        (() => Meteor.call('customer.new', 'a')).should.throw();
-        (() => Meteor.call('customer.new', [])).should.throw();
-        (() => Meteor.call('customer.new', true)).should.throw();
       });
 
       it('returns the id of the new customer', () => {
@@ -49,6 +42,35 @@ if (Meteor.isServer) {
         const id = Meteor.call('customer.new', {});
         Customers.findOne(id).shipments.should.be.instanceOf(Array);
         Customers.findOne(id).shipments.length.should.equal(0);
+      });
+    });
+
+    describe('customer.save', () => {
+      let customerId;
+      beforeEach(() => {
+        customerId = Customers.insert({
+          name: 'b',
+          address: 'c',
+          emailAddress: 'd',
+          currency: 'e',
+        });
+      });
+
+      it('saves changes to the customer', () => {
+        Meteor.call('customer.save', customerId, { name: '1' });
+        Customers.findOne(customerId).name.should.equal('1');
+        Meteor.call('customer.save', customerId, { address: '1' });
+        Customers.findOne(customerId).address.should.equal('1');
+        Meteor.call('customer.save', customerId, { emailAddress: '1' });
+        Customers.findOne(customerId).emailAddress.should.equal('1');
+        Meteor.call('customer.save', customerId, { currency: '1' });
+        Customers.findOne(customerId).currency.should.equal('1');
+      });
+
+      it('throws an error if other properties are passed', () => {
+        (() => Meteor.call('customer.save', customerId, { _id: 'x' }))
+          .should
+          .throw();
       });
     });
   });