+++ imports/state/reducers/reducers.js
@@ -1,2 +1,5 @@
 export { newQuote } from './quote/newQuoteReducers';
 export { customer } from './customer/customerReducers';
+export { branch } from './branch/branchReducers';
+export { customerList } from './customerList/customerListReducers';
+export { userProfile } from './userProfile/userProfileReducers';