+++ imports/ui/pages/Customer.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Route } from 'react-router-dom';
+import { Route, Link } from 'react-router-dom';
 import { createContainer } from 'meteor/react-meteor-data';
 import { Meteor } from 'meteor/meteor';
 
@@ -10,7 +10,7 @@ import QuoteListItem from '../list-items/QuoteListItem.jsx';
 import { Customers } from '../../api/customers/customersCollection';
 import { Quotes } from '../../api/quotes/quotesCollection';
 
-import EditCustomer from '../editors/EditCustomer.jsx';
+import EditCustomerConnect from '../editors/EditCustomerConnect';
 import EditQuoteHeaderConnect from '../editors/EditQuoteHeaderConnect.jsx';
 import EditQuoteChargesConnect from '../editors/EditQuoteChargesConnect';
 import EditQuoteEmailConnect from '../editors/EditQuoteEmailConnect';
@@ -20,6 +20,12 @@ const CustomerInner = ({ customer, loading, history }) => {
   return (
     <div className="">
       <div className="content customer">
+        <div className="process-header">
+          <div className="title">CUSTOMER</div>
+          <Link to="/customers">
+            <button className="button-primary">BACK TO CUSTOMER LIST</button>
+          </Link>
+        </div>
         {
           loading ?
             null :
@@ -54,10 +60,6 @@ const CustomerInner = ({ customer, loading, history }) => {
           }
         />
         <Route
-          path="/customers/:customerId/edit"
-          render={props => <EditCustomer {...props} />}
-        />
-        <Route
           path="/customers/view/:customerId/quotes/:quoteId/header"
           render={props => <EditQuoteHeaderConnect {...props} />}
         />
@@ -77,7 +79,7 @@ const CustomerInner = ({ customer, loading, history }) => {
       <div className="content-footer-accent customers-footer-accent" />
     </div>
   );
-}
+};
 
 CustomerInner.propTypes = {
   customer: PropTypes.object,